[programming_examples][ml/yolo26n] simplify; vectorization unlock; new make UX
Cleanup pass that also caught a Makefile rule conflict silently shipping
scalar reference kernels in place of the vectorized _vec.cc sources.
Fixing it drove a 14x speedup on m8 standalone and 1.8x on the chain
N=1 dispatch — all BIT-EXACT vs ORT.

Purge:
  - scripts/m8_megakernel.py (1-tile attempt, never worked in chain)
  - scripts/m8_stage.py (older 8-tile m8 design, superseded by megakernel)
  - aie2_yolo_iron.py (full-chain builder, redundant with iron_partial)
  - scripts/memtile_bw_bench.py + memtile_bw_run.py (microbench, unused)
  - 16 scalar reference .cc + 5 dead .h in kernels/ (each had a _vec.cc
    counterpart that the Makefile claimed to use but didn't)
  - M8_MODEL.md, MEGAKERNEL_M8_PLAN.md, FUSION_PLAN.md, VECTORIZATION_PLAN.md
    (historical plans with predictions long since superseded)

Wiring:
  - aie2_yolo_per_block.py: the 2-tile megakernel is now the only m8 path
    (was opt-in via M8_MEGAKERNEL_2TILE=1). Per-block + chain builders
    call it directly; no env-var dispatch.
  - scripts/m8_megakernel_2tile.py: strip debug env knobs (M8_NOKERNEL,
    M8_NK_*, M8_SKIP_COPY, M8_ACT_IN_DEPTH, M8_N_*_CHUNKS) added during
    Phase B perf attribution; keep TRACE_* plumbing.
  - kernels/yolo_{m8_*,c3k2_heavy_inner_pair_cv{1,2}_streamed}_vec.cc:
    cast power-of-2 divide dividends to uint32_t so Peano lowers to
    shifts instead of calling __divsi3.

Makefile:
  - Remove eval rules that overrode the _vec.cc rules with the scalar
    .cc rules (the rule-conflict bug behind the silently-scalar kernels),
    plus rules for now-deleted suffixed pair0/pair1 variants.
  - New targets: time, time_chain, trace, clean_chain, help.
  - CHAIN_N_SAMPLES stamp file forces chain MLIR rebuild when the value
    changes.
  - TIME_ARGS env passthrough to scripts/time_block.py + scripts/time_chain.py.

test_chain_ort.py: honor CHAIN_N_SAMPLES — feed N copies of the test
image and compare every output slice against the single ORT reference.

scripts/trace_m8.sh: new — build with TRACE_*, run, parse with
python/utils/trace/parse.py + get_trace_summary.py. Wired as `make trace`.

python/utils/trace/parse.py: init cycles = 0 so traces whose first
command is a Repeat parse without UnboundLocalError.

Documentation:
  - README: rewrite. Adds Performance section, Make targets reference,
    Tracing how-to, and a "Contributing & debugging" section with common
    patterns and recipes (L1 overflow, bit-exact failures, deadlock,
    __divsi3, reading the compiled .o for vector ops, etc.).
  - placement.py: header updated (26/32 tiles, m8 megakernel). m9
    description corrected to 7 tiles. PLACEMENT["m8"] now reflects the
    2-tile reality.
  - tile_layout.py: BLOCKS table + LAYOUT diagram + dataflow narrative
    updated to current placement. Removed stale CASCADES list and the
    "open design questions" section (all resolved).
  - scripts/m9_stage.py: stage map docstring corrected (1..10, with 10
    being the full PSA — was claiming "only stage 1 is wired").

Measurements (Strix HW, n=10-20 iters, std < 0.2 ms):

  Chain end-to-end vs ORT (BIT-EXACT, all samples):
    N=1   single-sample latency : 397.4 ms       (2.52 fps)
      was 715.8 ms                                (1.80x faster)
    N=15  batched throughput    : 318.5 ms/sample (3.14 fps)
      was 404.4 ms/sample                         (1.27x faster)

  Per-block standalone median ms (BIT-EXACT m0..m8 vs ORT):
    m0  55.4   m1  45.1   m2  26.4   m3  76.4   m4  22.0
    m5  73.6   m6  15.1   m7  36.5   m8  28.2
      m6 was ~280 ms (~18x faster), m8 was 406.7 ms (~14x faster)

  m3 / m5 (chunked conv_stride, 76 / 74 ms standalone) are the new
  chain critical-path stages.

+++ programming_examples/ml/yolo26n/kernels/yolo_c3k2_heavy_inner_pair_cv2_skip_streamed_vec.cc
@@ -65,17 +65,19 @@ void KERNEL_NAME(yolo_c3k2_heavy_inner_pair_cv2_skip_streamed_silu_bias_i8_i8)(
     const int32_t skip_rsh_add) {
   event0();
 
-  const int32_t chunk_oc = output_channels / n_chunks;
+  // See pair_cv1 _streamed_vec.cc: cast to unsigned so /power-of-2 lowers
+  // to shifts instead of calling __divsi3.
+  const int32_t chunk_oc = (uint32_t)output_channels / (uint32_t)n_chunks;
   const int32_t oc_offset = chunk_idx * chunk_oc;
   const bool skip_top = (border == 0);
   const bool skip_bot = (border == 2);
   const int ky_start = skip_top ? 1 : 0;
   const int ky_end = skip_bot ? 2 : 3;
 
-  const int ic_tiles = input_channels / 8;
-  const int chunk_oc_tiles = chunk_oc / 8;
+  const int ic_tiles = (uint32_t)input_channels / 8u;
+  const int chunk_oc_tiles = (uint32_t)chunk_oc / 8u;
   const int output_width = input_width;
-  const int x_tiles = output_width / 4;
+  const int x_tiles = (uint32_t)output_width / 4u;
 
   ::aie::set_saturation(aie::saturation_mode::saturate);
   ::aie::set_rounding(aie::rounding_mode::positive_inf);
